menus refactoring

+++ server/controllers/language/actionsLanguage.js
@@ -1,5 +1,5 @@
 const languageTable = require('../../models/index').languages;
-
+const validationError = require("../../services/validationError");
 module.exports = {
     getLanguage: async (languageId) => {
         try {
@@ -25,57 +25,85 @@ module.exports = {
             
         } catch (error) {
             console.log(error);
-            return {
-                message:error.message,
-                statusCode:204
-            }
+            return validationError(error)
         }
     },
     getLanguages: async () => {
         try {
-            return await languageTable.findAll({
+            const result= await languageTable.findAll({
                 attributes: {
                     exclude: ['createdAt', 'updatedAt'],
                 },
             });
+            return {
+                statusCode:200,
+                message:result
+            }
         } catch (error) {
             console.log(error);
-            return error;
+            return validationError(error);
         }
     },
     addLanguage: async (body) => {
         try {
-            return await languageTable.create({ ...body });
+            const result= await languageTable.create({ ...body });
+            return {
+                statusCode:200,
+                message:result
+            }
         } catch (error) {
             console.log(error);
-            return error;
+            return validationError(error);
         }
     },
     editLanguage: async (languageId, body) => {
         try {
-            return await languageTable.update(
-                { ...body },
-                {
-                    where: {
-                        id: languageId,
-                    },
+            const resultFind=await languageTable.findOne({where:{id:languageId}})
+            if(resultFind!==null){
+                await languageTable.update(
+                    { ...body },
+                    {
+                        where: {
+                            id: languageId,
+                        },
+                    }
+                );
+                return {
+                    statusCode:200,
+                    message:`Language id ${languageId} changed successfully`
                 }
-            );
+            }
+            return {
+                statusCode:401,
+                message:'Language with this id does not exist'
+            }
         } catch (error) {
             console.log(error);
-            return error;
+            return validationError(error);
         }
     },
     delLanguage: async (languageId) => {
         try {
-            return await languageTable.destroy({
-                where: {
-                    id: languageId,
-                },
-            });
+            const resultFind=await languageTable.findOne({where:{id:languageId}})
+            if(resultFind!==null){
+                await languageTable.destroy({
+                    where: {
+                        id: languageId,
+                    },
+                });
+                return {
+                    statusCode:200,
+                    message:`Language id ${languageId} removed successfully`
+                }
+            }
+            return {
+                statusCode:401,
+                message:'Language with this id does not exist'
+            }
+           
         } catch (error) {
             console.log(error);
-            return error;
+            return validationError(error);
         }
     },
 };
